+++ CHANGES.md
@@ -1,6 +1,13 @@
 <!-- markdownlint-configure-file { "no-duplicate-heading": { "siblings_only": true } } -->
 
 <!-- markdownlint-disable-next-line first-line-h1 -->
+
+## Unreleased
+
+### Added
+
+* `browser_defaults` to [CMarkStyle](CMarkStyle)
+
 ## 0.1.1 - 2026-05-01
 
 ### Fixed